+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 
 ## [0.2] - Unreleased
 
-- 
+- Faces `pgmacs-table-header` and `pgmacs-table-data` are used to display the header and the rows of
+  database tables.